Init randomizer and rename artfacts_module
