+++ master_thesis/generator/artifacts_generator.py
