+++ master_thesis/generator/randomizer.py
@@ -1,0 +1,5 @@
+
+
+class Randomiszer:
+    def __init__(self, seed: int= 23):
+        self.seed = seed
